fix(profile.component.ts): add check to prevent deletion of posts when no post is selected

+++ src/app/profile/profile.component.ts
@@ -279,6 +279,10 @@ export class ProfileComponent implements OnInit {
   }
 
   deleteMultiple() {
+    if(this.selected.length == 0) {
+      return;
+    }
+
     this._profileService.deletePost({ id: this.selected }).subscribe((res) => {
       if (res['code'] == 200) {
         this.selectedPosts = 0;
